Move general purpose type defintions into top-level src directory
