+++ src/utils/class-utils/getClassInheritance.ts
@@ -1,5 +1,5 @@
 import getSuperClass from "./getSuperClass";
-import { Class, ClassInstance } from "./types";
+import { Class, ClassInstance } from "../../types";
 
 /**
  * Retrieves an array of JavaScript classes which form the inheritance of the
